v0.3 - Moved functions from header.php
