+++ inc/wp-json.php
@@ -1,0 +1,45 @@
+<?php
+
+/*
+* Remove JSON API links in header html
+*/
+function myTheme_removeJSON_API() {
+
+  // Remove the REST API lines from the HTML Header
+  remove_action( 'wp_head', 'rest_output_link_wp_head', 10 );
+  remove_action( 'wp_head', 'wp_oembed_add_discovery_links', 10 );
+
+  // Remove the REST API endpoint.
+  remove_action( 'rest_api_init', 'wp_oembed_register_route' );
+
+  // Turn off oEmbed auto discovery.
+  add_filter( 'embed_oembed_discover', '__return_false' );
+
+  // Don't filter oEmbed results.
+  remove_filter( 'oembed_dataparse', 'wp_filter_oembed_result', 10 );
+
+  // Remove oEmbed discovery links.
+  remove_action( 'wp_head', 'wp_oembed_add_discovery_links' );
+
+  // Remove oEmbed-specific JavaScript from the front-end and back-end.
+  remove_action( 'wp_head', 'wp_oembed_add_host_js' );
+
+}
+add_action( 'after_setup_theme', 'myTheme_removeJSON_API' );
+
+/*
+	Snippet completely disable the REST API and shows {"code":"rest_disabled","message":"The REST API is disabled on this site."} 
+	when visiting http://yoursite.com/wp-json/
+*/
+function myTheme_disableJSON_API() {
+
+  // Filters for WP-API version 1.x
+  add_filter('json_enabled', '__return_false');
+  add_filter('json_jsonp_enabled', '__return_false');
+
+  // Filters for WP-API version 2.x
+  add_filter('rest_enabled', '__return_false');
+  add_filter('rest_jsonp_enabled', '__return_false');
+
+}
+add_action( 'after_setup_theme', 'myTheme_disableJSON_API' );
